refactor: 패키지 및 메서드 분리
