+++ src/main/java/lotto/Main.java
@@ -1,5 +1,8 @@
 package lotto;
 
+import lotto.view.InputView;
+import lotto.view.OutputView;
+
 public class Main {
 
     public static void main(String[] args) {
